+++ src/lib/config.ts
@@ -60,7 +60,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "latest",
-    sdkVersion: "0.6.3",
-    genVersion: "2.401.2",
-    userAgent: "speakeasy-sdk/typescript 0.6.3 2.401.2 latest @lukehagar/discoursejs",
+    sdkVersion: "0.7.0",
+    genVersion: "2.402.5",
+    userAgent: "speakeasy-sdk/typescript 0.7.0 2.402.5 latest @lukehagar/discoursejs",
 } as const;