+++ RELEASES.md
@@ -219,3 +219,13 @@ Based on:
 - [typescript v0.6.3] .
 ### Releases
 - [NPM v0.6.3] https://www.npmjs.com/package/@lukehagar/discoursejs/v/0.6.3 - .
+
+## 2024-08-20 00:22:26
+### Changes
+Based on:
+- OpenAPI Doc latest 
+- Speakeasy CLI 1.376.0 (2.402.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.0] .
+### Releases
+- [NPM v0.7.0] https://www.npmjs.com/package/@lukehagar/discoursejs/v/0.7.0 - .